generative_ai/requirements.txt
transformers
torch
openai
google-cloud-aiplatform
Pillow
langchain
requests
pandas
matplotlib
